:recycle: refactor: util 함수 따로 뺌
