+++ src/utils/format-date.ts
@@ -1,0 +1,9 @@
+const formatDate = (date: Date) => {
+  const yyyy = date.getFullYear();
+  const mm = String(date.getMonth() + 1).padStart(2, '0');
+  const dd = String(date.getDate()).padStart(2, '0');
+
+  return `${yyyy}.${mm}.${dd}`;
+};
+
+export default formatDate;
